update: ランキング全体をコンポーネント化
